moved _create_mrjob_zip() from base runner class into bin runner
